moving toBoolean to omero.gateway.utils as it is a common helper
