+++ components/tools/OmeroWeb/omeroweb/webadmin/webadmin_utils.py
@@ -37,24 +37,6 @@ def upgradeCheck():
         logger.info("Upgrade check was manually disabled.\n")
 
 
-def toBoolean(val):
-    """
-    Get the boolean value of the provided input.
-
-        If the value is a boolean return the value.
-        Otherwise check to see if the value is in
-        ["false", "f", "no", "n", "none", "0", "[]", "{}", "" ]
-        and returns True if value is not in the list
-    """
-
-    if val is True or val is False:
-        return val
-
-    falseItems = ["false", "f", "no", "n", "none", "0", "[]", "{}", ""]
-
-    return not str(val).strip().lower() in falseItems
-
-
 def removeUnaddressable(experimenters):
     """
     Remove experimenters that do not have email addresses or do not pass
